// YOUCOM.ReserVook/src/app/features/web/master/sitecontroller/model/sitecontroller.model.ts
import { Base } from '../../../../../shared/model/baseinfo.model'

export class SiteControllerInfo{
  companyNo: string;
  scCd: string;
  scVer: string;
  scUseFlg: string;
  scSystemId: string;
  scUsrId: string;
  scUsrPassword: string;
  scRcvCommMethod: string;
  scRcvCommFormat: string;
  scRcvCommVer: string;
  scRcvInterval: number;
  scNewRcvFlg: string;
  scChangeRcvFlg: string;
  scCancellationRcvFlg: string;
  scReservationRcvUrl: string;
  scReservationRcvCompUrl: string;
  scRcvFilePath: string;
  scRcvCompFilePath: string;
  scRcvErrorFilePath: string;
  scRcvFilePattern: string;
  scRcvFileExt: string;
  scRcvFilePrsvTerm: number;
  scRcvDataPrsvTerm: number;
  scRcvDefBlockCd: string;
  scRcvClerkCd: string;
  scSndInterval: number;
  scRmngRmsSndFlg: string;
  scSndCommMethod: string;
  scSndCommFormat: string;
  scSndCommVer: string;
  scRmngRmsSndUrl: string;
  scSndFilePath: string;
  scSndDataPrsvTrem: number;
  scRsvNoNameAdd: string;
  updateCnt: number;
  programId: string;
  createClerk: string;
  createMachineNo: string;
  createMachine: string;
  createDatetime: string;
  updateClerk: string;
  updateMachineNo: string;
  updateMachine: string;
  updateDatetime: string;

  cdName: string;
  dispScUseFlg: string;
  dispScNewRcvFlg: string;
  dispScCancellationRcvFlg: string;

}